separe call api from component
