+++ src/components/SearchField.tsx
@@ -1,58 +1,6 @@
 import React, { useState, useEffect, useContext } from 'react';
-import axios from 'axios';
 import { TokenContext } from '../TokenContext';
-
-interface ItemPush {
-  title: string;
-  artist: string;
-  url: string;
-  image: string;
-}
-
-function makeObjectOfDataRequest(data): Array<ItemPush> {
-  const dataTracks: Array<ItemPush> = [];
-
-  data.tracks.items.forEach((item: Record<string, any>) => {
-    const i: ItemPush = {
-      title: item.name,
-      url: item.external_urls.spotify,
-      image:
-        item.album.images !== undefined && item.album.images.length > 0 ? item.album.images[0].url : 'noImageFound',
-      artist: item.artists.name,
-    };
-    dataTracks.push(i);
-  });
-  return dataTracks;
-}
-
-async function searchOnSpotify(token, word = 'julien'): Promise<Array<ItemPush>> {
-  if (word === '') return [];
-
-  const spotifyToken = token;
-  const url = 'https://api.spotify.com/v1/search';
-  const config = {
-    params: {
-      q: word,
-      type: 'track',
-    },
-    headers: {
-      Authorization: `Bearer ${spotifyToken}`,
-    },
-  };
-
-  return axios
-    .get(url, config)
-    .then((response) => response.data)
-    .then((data) => {
-      return makeObjectOfDataRequest(data);
-    })
-    .catch((e) => {
-      if (e.response.status !== 200) {
-        alert(e.response.data.error.message);
-      }
-      return [];
-    });
-}
+import { ItemPush, searchOnSpotify } from './RequestSpotify';
 
 export default function SearchField(): JSX.Element {
   const { services } = useContext(TokenContext);
